Do not use multiple techniques to start and stop the stream
Returning paComplete from the audio callback, as well as using our own stop variable may conflict in complex ways. One such example is if the stop boolean is set and the device is required to read the nullptr buffer when the audio FIFO thread is closing. Also, PortAudio's says using paComplete and paAbort is an alternative technique to calling Pa_StopStream, so using both isn't necessary and thus avoids streams being stopped or started unexpectedly.