Remove locks in Track destructor
If there is contention, where one thread wants to do something with the track and the other is trying to destroy it, we are betting that the destructor will win over the lock last, which wont always happen. The real fix in this instance is to ensure that when the destructor is called, the object is no longer in use by anything else.

By removing the lock, this isn't undoing a problem. The problem was always still there. The lock will only help 50% of the time (i.e., in the scenario mentioned above), with the other 50% crashing the program. The chances get slimmer as the contention increases, but this doesn't change the fact that this is still error-prone.