Use direct connections when adding track or clip views
QueuedConnections were used previously. Because this kind of connection only invokes the slot when Qt returns to the event loop, this could be a problem if the track or clip was deleted in between the time when the signal that a track or clip was added was emitted, and when the slot to add the corresponding track or clip view was invoked.

An alternative solution that could be even better overall is to use weak pointers in the view classes and manage the objects using shared pointers. This is a more involved task, however.